+++ owsi-core/owsi-core-components/owsi-core-component-jpa-more/src/main/java/fr/openwide/core/jpa/more/business/audit/service/IAuditService.java
@@ -25,7 +25,7 @@ import fr.openwide.core.jpa.more.business.audit.model.AbstractAudit;
 /**
  * <p>Service concernant les {@link AbstractAudit}.</p>
  */
-public interface IAuditService<T extends AbstractAudit> extends IGenericEntityService<Integer, T> {
+public interface IAuditService<T extends AbstractAudit> extends IGenericEntityService<Long, T> {
 
 	/**
 	 * Renvoie le sujet d'un Audit.
