+++ owsi-core/owsi-core-components/owsi-core-component-jpa/src/main/java/fr/openwide/core/jpa/business/generic/util/GenericEntityIdComparator.java
@@ -5,14 +5,14 @@ import java.util.Comparator;
 
 import fr.openwide.core.jpa.business.generic.model.GenericEntity;
 
-public class GenericEntityIdComparator implements Comparator<GenericEntity<Integer, ?>>, Serializable {
+public class GenericEntityIdComparator implements Comparator<GenericEntity<Long, ?>>, Serializable {
 	
 	private static final long serialVersionUID = -9178542049081510289L;
 
-	public static final Comparator<GenericEntity<Integer, ?>> INSTANCE = new GenericEntityIdComparator();
+	public static final Comparator<GenericEntity<Long, ?>> INSTANCE = new GenericEntityIdComparator();
 
 	@Override
-	public int compare(GenericEntity<Integer, ?> o1, GenericEntity<Integer, ?> o2) {
+	public int compare(GenericEntity<Long, ?> o1, GenericEntity<Long, ?> o2) {
 		if (o1.getId() != null && o2.getId() != null) {
 			return o1.getId().compareTo(o2.getId());
 		} else if (o1.getId() != null && o2.getId() == null) {
