+++ owsi-core/owsi-core-components/owsi-core-component-jpa/src/test/java/fr/openwide/core/test/jpa/example/business/person/service/PersonService.java
@@ -22,7 +22,7 @@ import javax.persistence.metamodel.SingularAttribute;
 import fr.openwide.core.jpa.business.generic.service.IGenericEntityService;
 import fr.openwide.core.test.jpa.example.business.person.model.Person;
 
-public interface PersonService extends IGenericEntityService<Integer, Person> {
+public interface PersonService extends IGenericEntityService<Long, Person> {
 
 	Long count(SingularAttribute<? super Person, String> attribute, String value);
 }
