+++ owsi-core/owsi-core-examples/owsi-core-example-jpa/src/main/java/fr/openwide/jpa/example/business/company/service/CompanyService.java
@@ -3,7 +3,7 @@ package fr.openwide.jpa.example.business.company.service;
 import fr.openwide.core.jpa.business.generic.service.IGenericEntityService;
 import fr.openwide.jpa.example.business.company.model.Company;
 
-public interface CompanyService extends IGenericEntityService<Integer, Company> {
+public interface CompanyService extends IGenericEntityService<Long, Company> {
 
 //	void addEmployee(Company company, Person employee) throws ServiceException, SecurityServiceException;
 //
