+++ owsi-core/owsi-core-examples/owsi-core-example-jpa/src/main/java/fr/openwide/jpa/example/business/person/dao/PersonDao.java
@@ -3,6 +3,6 @@ package fr.openwide.jpa.example.business.person.dao;
 import fr.openwide.core.jpa.business.generic.dao.IGenericEntityDao;
 import fr.openwide.jpa.example.business.person.model.Person;
 
-public interface PersonDao extends IGenericEntityDao<Integer, Person> {
+public interface PersonDao extends IGenericEntityDao<Long, Person> {
 
 }
