+++ owsi-core/owsi-core-examples/owsi-core-example-jpa/src/main/java/fr/openwide/jpa/example/business/project/dao/ProjectDao.java
@@ -3,6 +3,6 @@ package fr.openwide.jpa.example.business.project.dao;
 import fr.openwide.core.jpa.business.generic.dao.IGenericEntityDao;
 import fr.openwide.jpa.example.business.project.model.Project;
 
-public interface ProjectDao extends IGenericEntityDao<Integer, Project> {
+public interface ProjectDao extends IGenericEntityDao<Long, Project> {
 
 }
