+++ owsi-core/owsi-core-examples/owsi-core-example-jpa/src/main/java/fr/openwide/jpa/example/business/project/service/ProjectService.java
@@ -3,6 +3,6 @@ package fr.openwide.jpa.example.business.project.service;
 import fr.openwide.core.jpa.business.generic.service.IGenericEntityService;
 import fr.openwide.jpa.example.business.project.model.Project;
 
-public interface ProjectService extends IGenericEntityService<Integer, Project> {
+public interface ProjectService extends IGenericEntityService<Long, Project> {
 
 }
